+++ Software/e-Agro/RadnaMjesta.cs
@@ -23,7 +23,10 @@ namespace e_Agro
             using (var context = new PI20_024_DBEntities())
             {
                 context.radno_mjesto.Attach(radnoMjesto);
-                context.radno_mjesto.Remove(radnoMjesto);
+                if(radnoMjesto.korisniks.Count > 0)
+                    System.Windows.Forms.MessageBox.Show("Ne možete obrisati radno mjesto koje ima zaposlenike!");
+                else
+                    context.radno_mjesto.Remove(radnoMjesto);
                 context.SaveChanges();
             }
         }
